Add a method for converting ZStream[Byte] to a ZManaged[java.io.InputStream]
not extension methods
