+++ streams/shared/src/main/scala/zio/stream/StreamEffectChunk.scala
@@ -135,6 +135,33 @@ private[stream] class StreamEffectChunk[-R, +E, +A](override val chunks: StreamE
         }
       }
     }
+
+  override final def toInputStream(
+    implicit ev0: E <:< Throwable,
+    ev1: A <:< Byte
+  ): ZManaged[R, E, java.io.InputStream] =
+    for {
+      thunk <- chunks.processEffect
+      javaStream = {
+        new java.io.InputStream {
+          var counter            = 0
+          var chunk: Chunk[Byte] = Chunk.empty
+          override def read(): Int =
+            try {
+              while (counter >= chunk.length) {
+                chunk = thunk().asInstanceOf[Chunk[Byte]]
+                counter = 0
+              }
+              val item = chunk(counter).toInt
+              counter += 1
+              item
+            } catch {
+              case StreamEffect.End        => -1
+              case StreamEffect.Failure(e) => throw e.asInstanceOf[E]
+            }
+        }
+      }
+    } yield javaStream
 }
 
 private[stream] object StreamEffectChunk extends Serializable {
